Add files via upload
These are the codes for minimizing the cost by taking optimal MW from different power stations, which are having different constraints on minimum and maximum power,impacts from fuel cost and running cost. This is a optimisation problem with various constraints solved using bio inspired optimisation.

The logic is developed by me but the concept and outline is taken from elsewhere.